refactor(play-records): #2437-2 extract PlayRecordDetailBody (prop-driven)
